requirements.txt
grpcio
grpcio-tools
pyyaml
tabulate